Change the exposed avocADO instance in catseffect3 for the generic one
