+++ cats-effect-3/src/main/scala/catseffect3.scala
@@ -1,11 +1,12 @@
 package avocado.catseffect3
 
 import avocado.AvocADO
-import _root_.cats.effect.IO
+import cats.Monad
+import cats.effect.kernel.Async
 
-given AvocADO[IO] = new AvocADO[IO] {
-  def pure[A](a: A): IO[A] = IO(a)
-  def map[A, B](fa: IO[A], f: A => B): IO[B] = fa.map(f)
-  def zip[A, B](fa: IO[A], fb: IO[B]): IO[(A, B)] = fa.both(fb)
-  def flatMap[A, B](fa: IO[A], f: A => IO[B]): IO[B] = fa.flatMap(f)
+given [F[_]: Monad: Async] : AvocADO[F] = new AvocADO[F] {
+  def pure[A](a: A): F[A] = Monad[F].pure(a)
+  def map[A, B](fa: F[A], f: A => B): F[B] = Monad[F].map(fa)(f)
+  def zip[A, B](fa: F[A], fb: F[B]): F[(A, B)] = Async[F].both(fa, fb)
+  def flatMap[A, B](fa: F[A], f: A => F[B]): F[B] = Monad[F].flatMap(fa)(f)
 }
