+++ cats-effect-3/src/test/scala/CatsEffectTCTests.scala
@@ -1,7 +1,7 @@
 package avocado.tests
 
 import avocado.*
-// import avocado.catseffect3.given
+import avocado.catseffect3.given
 
 import scala.concurrent.duration.*
 import cats.Monad
@@ -11,13 +11,6 @@ import cats.effect.IO
 import cats.effect.unsafe.implicits.*
 
 class CatsEffectTCTests extends munit.FunSuite {
-
-  given [F[_]: Monad: Async] : AvocADO[F] = new AvocADO[F] {
-    def pure[A](a: A): F[A] = Monad[F].pure(a)
-    def map[A, B](fa: F[A], f: A => B): F[B] = Monad[F].map(fa)(f)
-    def zip[A, B](fa: F[A], fb: F[B]): F[(A, B)] = Async[F].both(fa, fb)
-    def flatMap[A, B](fa: F[A], f: A => F[B]): F[B] = Monad[F].flatMap(fa)(f)
-  }
 
   def testWithTimeLimit(name: String, maxMillis: Long)(body: => Unit): Unit = {
     test(name) {
